+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXModule.java
@@ -249,6 +249,7 @@ public class AppLovinMAXModule
 
                 WritableMap sdkConfiguration = Arguments.createMap();
                 sdkConfiguration.putInt( "consentDialogState", configuration.getConsentDialogState().ordinal() );
+                sdkConfiguration.putString( "countryCode", configuration.getCountryCode() );
                 callback.invoke( sdkConfiguration );
             }
         } );